school/requirements.txt
asgiref==3.3.1
Django==2.2
django-fontawesome==1.0
django-widget-tweaks==1.4.8
image==1.5.33
Pillow==8.1.0
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
